Adapt canopen::openConnection to Adlink driver
We reduced the range of available baudrate, since our card
supports values down to 125Kbps. Moreover, we added a #DEFINE to
match human-readable representation of the baudrate values with the
internal coding used by the Adlink driver (integers from 0 to 3).

The configuration used for the port is the following:
    - Standard frame (11-bit). The same used by libpcan with the
      CAN_INIT_TYPE_ST flag.
    - No input filter based on the MAC_ID.
    - Baudrate passed as parameter.

The compatibility problem for this function is that with libpcan
the port is identified by a string, while Adlink uses the indexes
of the card and of the port.